+++ packages/firebase_database/lib/src/query.dart
@@ -51,6 +51,7 @@ class Query {
           'Query#observe',
           <String, dynamic>{
             'app': _database.app?.name,
+            'databaseURL': _database.databaseURL,
             'path': path,
             'parameters': _parameters,
             'eventType': eventType.toString(),
@@ -66,6 +67,7 @@ class Query {
             'Query#removeObserver',
             <String, dynamic>{
               'app': _database.app?.name,
+              'databaseURL': _database.databaseURL,
               'path': path,
               'parameters': _parameters,
               'handle': handle,
@@ -102,7 +104,11 @@ class Query {
   /// than or equal to the given key.
   Query startAt(dynamic value, {String key}) {
     assert(!_parameters.containsKey('startAt'));
-    assert(value is String || value is bool || value is double || value is int);
+    assert(value is String ||
+        value is bool ||
+        value is double ||
+        value is int ||
+        value == null);
     final Map<String, dynamic> parameters = <String, dynamic>{'startAt': value};
     if (key != null) parameters['startAtKey'] = key;
     return _copyWithParameters(parameters);
@@ -114,7 +120,11 @@ class Query {
   /// than or equal to the given key.
   Query endAt(dynamic value, {String key}) {
     assert(!_parameters.containsKey('endAt'));
-    assert(value is String || value is bool || value is double || value is int);
+    assert(value is String ||
+        value is bool ||
+        value is double ||
+        value is int ||
+        value == null);
     final Map<String, dynamic> parameters = <String, dynamic>{'endAt': value};
     if (key != null) parameters['endAtKey'] = key;
     return _copyWithParameters(parameters);
@@ -126,7 +136,11 @@ class Query {
   /// If a key is provided, there is at most one such child as names are unique.
   Query equalTo(dynamic value, {String key}) {
     assert(!_parameters.containsKey('equalTo'));
-    assert(value is String || value is bool || value is double || value is int);
+    assert(value is String ||
+        value is bool ||
+        value is double ||
+        value is int ||
+        value == null);
     return _copyWithParameters(
       <String, dynamic>{'equalTo': value, 'equalToKey': key},
     );
@@ -196,6 +210,7 @@ class Query {
       'Query#keepSynced',
       <String, dynamic>{
         'app': _database.app?.name,
+        'databaseURL': _database.databaseURL,
         'path': path,
         'parameters': _parameters,
         'value': value