+++ packages/package_info/ios/Classes/PackageInfoPlugin.m
@@ -18,6 +18,8 @@
     result([[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleShortVersionString"]);
   } else if ([call.method isEqualToString:@"getBuildNumber"]) {
     result([[NSBundle mainBundle] objectForInfoDictionaryKey:@"CFBundleVersion"]);
+  } else if ([call.method isEqualToString:@"getPackageName"]) {
+    result([[NSBundle mainBundle] bundleIdentifier]);
   } else {
     result(FlutterMethodNotImplemented);
   }