+++ packages/firebase_admob/CHANGELOG.md
@@ -4,6 +4,11 @@
 * **Breaking change**. The properties and parameters named "unitId" in BannerAd
   and InterstitialAd have been renamed to "adUnitId" to better match AdMob's
   documentation and UI.
+  
+## 0.2.3
+
+* Simplified and upgraded Android project template to Android SDK 27.
+* Updated package description.
 
 ## 0.2.2